Added test_util_tests, modified test_util, fixed time_util bugs

+++ lib/utils/datetime.hpp
@@ -6,15 +6,15 @@
 namespace TimeUtil {
 
     struct DateTime {
-        uint8_t     day = 0;
-        uint8_t     month = 0;
-        uint16_t    year = 0;
-        uint8_t     hours = 0;
-        uint8_t     minutes = 0;
-        uint8_t     seconds = 0;
-        int8_t      timezone_sign = 1;
-        int8_t      timezone_hours = 0;
-        int8_t      timezone_minutes = 0;
+        int day = 0;
+        int month = 0;
+        int year = 0;
+        int hours = 0;
+        int minutes = 0;
+        int seconds = 0;
+        int timezone_sign = 1;
+        int timezone_hours = 0;
+        int timezone_minutes = 0;
     };
 
 } // namespace TimeUtil
